scripts/pam-start-usertree: Run s6-rc again on >=2nd login
This is for the case where the last session is a tty session and the new
session is a x11/wayland session, which probably needs more services to
be running. These additions will make sure those needed services are
started even when the x11/wayland session is not the first session. This
also applies to other session types which may have different set of
services that need to be started.