pipewire-srv: Improve readiness handling
PipeWire 0.3.57 introduced PIPEWIRE_NOTIFICATION_FD to print the socket
address to the fd specified by that environment variable when the socket
is ready, which we can use to notify readiness to s6 by setting
PIPEWIRE_NOTIFICATION_FD to the same value as the content of
notification-fd file.

This environment variable is only used by the pipewire daemon though
(pipewire-media-session and pipewire-pulse don't use it).