update-sbo: Further refinements to gnome.org handlings
gnome_get_latest_release() will now compare major version first then the
minor version. It will also run on a subshell (using "()" instead of
"{}" to enclose the function body) so that the variable $MAJOR is not
saved in the main shell process. Major version substitution are also
performed on *.info file so that the newest version dowload link will be
correct (since there is /<software_name>/<major_version>/ part on the
download link where the <major_version> part also need to be
substituted).